requirements.txt
gradio
onnxruntime
mediapipe
protobuf
torch
torchvision
pillow
opencv-python-headless
numpy